Cleaning up some compiler warnings.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@508 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ libhb/hb.c
@@ -324,7 +324,7 @@ void hb_get_preview( hb_handle_t * h, hb_title_t * title, int picture,
     memset( filename, 0, 1024 );
 
     hb_get_tempory_filename( h, filename, "%x%d",
-                             (int) title, picture );
+                             (intptr_t) title, picture );
 
     file = fopen( filename, "r" );
     if( !file )
